feat: denort binary
